refector gnb in footer
